Status and Title have been moved under Toolbar
