+++ src/client/components/PageLayout/Toolbar/Status/StatusConfirm.scss
@@ -38,7 +38,5 @@
 
   &:hover {
     cursor: pointer;
-
   }
 }
-
